fix!: change default to "not set" for confirm and close
This fixes the issue where the Enter key is sent twice to the target
(first key event captured by the Obsidian app, then sent again from this
plugin). This issue is observable when `.app-container` is set as the
target, and `Enter` is registered as the confirm shortcut. If you press
enter in a normal area, you will see two newlines added.